+++ src/Events/ImportFailed.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace EightyNine\ExcelImport\Events;
+
+use Throwable;
+
+class ImportFailed
+{
+    public function __construct(
+        public readonly object $import,
+        public readonly Throwable $exception,
+    ) {}
+}