+++ src/Events/ImportStarted.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace EightyNine\ExcelImport\Events;
+
+class ImportStarted
+{
+    public function __construct(
+        public readonly object $import,
+    ) {}
+}